Add mutative mode
It is possible that prototype delegation could cause odd differences in
behaviour depending on who is accessing which object reference. Some code
may be referencing the source object, and some the target.

This isn't often a problem but can be, so we allow a mutative mode so
we can mutate the object directly and affect the source object.